refactoring packages to their specific subdomains
