--- src/main/java/br/com/caelum/vraptor/simplevalidator/ValidationStrategy.java
@@ -1,5 +1,0 @@
-package br.com.caelum.vraptor.simplevalidator;
-
-public interface ValidationStrategy<T> {
-
-}
