+++ src/main/java/br/com/caelum/vraptor/simplevalidator/strategy/CustomValidationStrategy.java
@@ -1,4 +1,5 @@
-package br.com.caelum.vraptor.simplevalidator;
+package br.com.caelum.vraptor.simplevalidator.strategy;
+
 
 public interface CustomValidationStrategy<T> extends ValidationStrategy<T>{
 	public void addErrors(T t);
