+++ src/main/java/br/com/caelum/vraptor/simplevalidator/strategy/ValidationStrategy.java
@@ -1,0 +1,5 @@
+package br.com.caelum.vraptor.simplevalidator.strategy;
+
+public interface ValidationStrategy<T> {
+
+}
