+++ src/main/java/br/com/caelum/vraptor/simplevalidator/strategy/ValidationStrategyHelper.java
@@ -1,4 +1,4 @@
-package br.com.caelum.vraptor.simplevalidator;
+package br.com.caelum.vraptor.simplevalidator.strategy;
 
 public interface ValidationStrategyHelper {
 	void addError(String message, Object...parameters);
